feat(tests): Move tests for `Git::new()` to `tests/`
These tests for `Git::new()` in `adm::vcs` module are not unit tests, but
integration tests. Thus, they belong in the `tests/` directory.
